requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
cryptography>=42.0
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
